+++ DoViBaker/DoViBaker.vcxproj
@@ -30,19 +30,19 @@
     <ClCompile Include="AvisynthEntry.cpp" />
     <ClCompile Include="DoViBaker.cpp" />
     <ClCompile Include="DoViCubes.cpp" />
-    <ClCompile Include="DoViTonemap.cpp" />
+    <ClCompile Include="DoViTransferFunctions.cpp" />
     <ClCompile Include="DoViMaxPqFileReader.cpp" />
     <ClCompile Include="DoViProcessor.cpp" />
-    <ClCompile Include="DoViTonemapper.cpp" />
+    <ClCompile Include="DoViTonemap.cpp" />
   </ItemGroup>
   <ItemGroup>
     <ClInclude Include="..\include\DoViBaker.h" />
     <ClInclude Include="..\include\DoViProcessor.h" />
     <ClInclude Include="..\include\dovi\rpu_parser.h" />
     <ClInclude Include="DoViCubes.h" />
-    <ClInclude Include="DoViTonemap.h" />
+    <ClInclude Include="DoViTransferFunctions.h" />
     <ClInclude Include="DoViMaxPqFileReader.h" />
-    <ClInclude Include="DoViTonemapper.h" />
+    <ClInclude Include="DoViTonemap.h" />
   </ItemGroup>
   <PropertyGroup Label="Globals">
     <VCProjectVersion>15.0</VCProjectVersion>
