+++ DoViBaker/DoViProcessor.cpp
@@ -3,7 +3,7 @@
 #include <string>
 
 #include "DoViProcessor.h"
-#include "DoViTonemap.h"
+#include "DoViTransferFunctions.h"
 
 
 DoViProcessor::DoViProcessor(const char* rpuPath, IScriptEnvironment* env, uint8_t blContainerBits, uint8_t elContainerBits)
@@ -205,10 +205,10 @@ bool DoViProcessor::intializeFrame(int frame, IScriptEnvironment* env, const uin
 
 		dynamic_min_pq = vdr_dm_data->dm_data.level1->min_pq;
 		dynamic_max_pq = vdr_dm_data->dm_data.level1->max_pq;
-		dynamic_max_content_light_level = DoViTonemap::pq2nits(dynamic_max_pq) + 0.5;
+		dynamic_max_content_light_level = DoViTransferFunctions::pq2nits(dynamic_max_pq) + 0.5;
 		if (vdr_dm_data->dm_data.level6) {
 			static_max_content_light_level = vdr_dm_data->dm_data.level6->max_content_light_level;
-			static_max_pq = DoViTonemap::nits2pq(vdr_dm_data->dm_data.level6->max_content_light_level);
+			static_max_pq = DoViTransferFunctions::nits2pq(vdr_dm_data->dm_data.level6->max_content_light_level);
 		}
 
 		skipTrim = true;
@@ -452,9 +452,9 @@ uint16_t DoViProcessor::signalReconstruction(uint16_t v, int16_t r) const {
 }
 
 void DoViProcessor::prepareTrimCoef() {
-	float x1 = trim.minNits = DoViTonemap::pq2nits(dynamic_min_pq);
-	float x2 = DoViTonemap::pq2nits(dynamic_avg_pq);
-	float x3 = trim.maxNits = DoViTonemap::pq2nits(dynamic_max_pq);
+	float x1 = trim.minNits = DoViTransferFunctions::pq2nits(dynamic_min_pq);
+	float x2 = DoViTransferFunctions::pq2nits(dynamic_avg_pq);
+	float x3 = trim.maxNits = DoViTransferFunctions::pq2nits(dynamic_max_pq);
 
 	float y1 = targetMinNits;
 	float y2 = sqrtf(x2 * sqrtf(targetMaxNits * targetMinNits));
@@ -480,18 +480,18 @@ void DoViProcessor::prepareTrimCoef() {
 }
 
 void DoViProcessor::processTrim(uint16_t& ro, uint16_t& go, uint16_t& bo, const uint16_t& ri, const uint16_t& gi, const uint16_t& bi) const  {
-	float dr = DoViTonemap::pq2nits(ri >> (outContainerBitDepth - out_bit_depth));
-	float dg = DoViTonemap::pq2nits(gi >> (outContainerBitDepth - out_bit_depth));
-	float db = DoViTonemap::pq2nits(bi >> (outContainerBitDepth - out_bit_depth));
+	float dr = DoViTransferFunctions::pq2nits(ri >> (outContainerBitDepth - out_bit_depth));
+	float dg = DoViTransferFunctions::pq2nits(gi >> (outContainerBitDepth - out_bit_depth));
+	float db = DoViTransferFunctions::pq2nits(bi >> (outContainerBitDepth - out_bit_depth));
 
 	float er = (trim.ccc[0] + dr * trim.ccc[1]) / (1 + dr * trim.ccc[2]);
 	float eg = (trim.ccc[0] + dg * trim.ccc[1]) / (1 + dg * trim.ccc[2]);
 	float eb = (trim.ccc[0] + db * trim.ccc[1]) / (1 + db * trim.ccc[2]);
 
 	if (trimInfoMissing) {
-		ro = DoViTonemap::nits2pq(er) << (outContainerBitDepth - out_bit_depth);
-		go = DoViTonemap::nits2pq(eg) << (outContainerBitDepth - out_bit_depth);
-		bo = DoViTonemap::nits2pq(eb) << (outContainerBitDepth - out_bit_depth);
+		ro = DoViTransferFunctions::nits2pq(er) << (outContainerBitDepth - out_bit_depth);
+		go = DoViTransferFunctions::nits2pq(eg) << (outContainerBitDepth - out_bit_depth);
+		bo = DoViTransferFunctions::nits2pq(eb) << (outContainerBitDepth - out_bit_depth);
 	}	else {
 		float y3 = targetMaxNits;
 		float fr = powf((std::clamp(((er / y3) * trim.goP[0]) + trim.goP[1], 0.0f, 1.0f)), trim.goP[2]) * y3;
@@ -503,8 +503,8 @@ void DoViProcessor::processTrim(uint16_t& ro, uint16_t& go, uint16_t& bo, const 
 		float gg = fg * powf((1 + trim.cS[0]) * fg / Y, trim.cS[1]);
 		float gb = fb * powf((1 + trim.cS[0]) * fb / Y, trim.cS[1]);
 
-		ro = DoViTonemap::nits2pq(gr) << (outContainerBitDepth - out_bit_depth);
-		go = DoViTonemap::nits2pq(gg) << (outContainerBitDepth - out_bit_depth);
-		bo = DoViTonemap::nits2pq(gb) << (outContainerBitDepth - out_bit_depth);
+		ro = DoViTransferFunctions::nits2pq(gr) << (outContainerBitDepth - out_bit_depth);
+		go = DoViTransferFunctions::nits2pq(gg) << (outContainerBitDepth - out_bit_depth);
+		bo = DoViTransferFunctions::nits2pq(gb) << (outContainerBitDepth - out_bit_depth);
 	}
 }
