+++ scripts/example.avs
@@ -5,7 +5,7 @@ el=FFVideoSource("clip.ts", threads=1, track=1)
 bl=bl.z_ConvertFormat(chromaloc_op="top_left=>top_left",width=bl.Width()/2,height=bl.Height()/2,resample_filter="spline36",resample_filter_uv="spline36")
 
 DoViBaker(bl,el)
-DoViTonemapper(masterMaxNits=-1, masterMinNits=0, targetMaxNits=1000, targetMinNits=0, lumScale=1.0).AVSCube("PQ2HLG.cube")
+DoViTonemap(masterMaxNits=-1, masterMinNits=0, targetMaxNits=1000, targetMinNits=0, lumScale=1.0).AVSCube("PQ2HLG.cube")
 z_ConvertFormat(chromaloc_op="center=>top_left",pixel_type="YUV420P16",colorspace_op="rgb:std-b67:2020:full=>2020ncl:std-b67:2020:limited",dither_type="none",resample_filter="spline36",resample_filter_uv="spline36")
 
 ScriptClip("""
